Shrink comment, move test a bit
